+++ CHANGELOG.md
@@ -1,5 +1,3 @@
-<!--todo: release initial version -->
-
 # Changelog
 
 All notable changes to this project will be documented in this file.
@@ -13,6 +11,20 @@ and this project adheres to [Dart Package Versioning](https://dart.dev/tools/pub
 
 - Base64 and Base64Url encoding schemes —
   [12](https://github.com/dartoos-dev/dartoos/issues/12)
+- Base64Decode class for decoding base64-encoded text —
+  [19](https://github.com/dartoos-dev/dartoos/issues/19).
+- Base64Norm class for normalizing base64-encoded texts.
+- Dartoos base64-encoding/decoding benchmark executable in 'example'.
+
+### Changed
+
+- Rand class accept as source of characters both plain and Future values.
+- BytesOf constructors accept both plain and Future values.
+
+### Removed
+
+- Rand.str constructor — **Breaking Change**.
+- FutureWrap.value constructor — **Breaking Change**.
 
 ## [0.1.0] - 2021-10-05